Display Update
This adds the ability to display the initial state of the game. 

An images folder is included for the images used when drawing.

graphics.py contains the graphics library used.

display.py handles displaying the game.

block,py had two functions added to get the width and height of a block.  The file name for each block was also calculated there.  self.size was also not being set, which was fixed.

unblockme.py had display-related additions made.  The portion of code that takes the board state and creates block objects was rewritten as it did not work at the time I copied that file.